Update extended_client.py
Extension of the echo application implemented in Part I to measure the round trip time (RTT) and throughput of the path connecting the client to the server.
The client repeatedly sends measurement messages to the server in an attempt to compute the mean RTT and/or mean throughput. A sample measurement is taken for each probe sent out.